pushing last feature developed, viewchatdetailpage.razor, razor.cs, razor.css. and many other etc.

+++ MauiApp1/Areas/Dashboard/ViewModel/CarbohydratesIntakeViewItem.cs
@@ -9,7 +9,7 @@ namespace MauiApp1.Areas.Dashboard.ViewModel
     public class CarbohydratesIntakeViewItem
     {
         public Guid CarbohydratesDataId { get; set; }
-        public long CarbohydratesIntakeCount { get; set; }
+        public double CarbohydratesIntakeCount { get; set; }
         public DateTimeOffset TransactionDate { get; set; }
     }
 }
